Add BigbeatObject
