+++ osu.Game.Rulesets.tau/Objects/TauBigBeatObject.cs
@@ -1,0 +1,10 @@
+﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+namespace osu.Game.Rulesets.Tau.Objects
+{
+    public class TauBigBeatObject : TauHitObject
+    {
+        //TODO: Think if Big Beats should give two combos instead.
+    }
+}
