+++ osu.Game.Rulesets.tau/UI/tauDrawableRuleset.cs
@@ -28,7 +28,19 @@ namespace osu.Game.Rulesets.Tau.UI
 
         protected override ReplayInputHandler CreateReplayInputHandler(Replay replay) => new TauFramedReplayInputHandler(replay);
 
-        public override DrawableHitObject<TauHitObject> CreateDrawableRepresentation(TauHitObject h) => new DrawabletauHitObject(h);
+        public override DrawableHitObject<TauHitObject> CreateDrawableRepresentation(TauHitObject h)
+        {
+            switch (h)
+            {
+                case TauBeatObject beatObject:
+                    return new DrawabletauBeatObject(beatObject);
+
+                case TauBigBeatObject bigBeat:
+                    return new DrawableTauBigBeatObject(bigBeat);
+            }
+
+            return null;
+        }
 
         protected override PassThroughInputManager CreateInputManager() => new TauInputManager(Ruleset?.RulesetInfo);
