+++ osu.Game.Rulesets.tau/osu.Game.Rulesets.tau.csproj
@@ -1,21 +1,21 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup Label="Project">
-    <TargetFramework>netcoreapp3.0</TargetFramework>
-    <AssemblyTitle>osu.Game.Rulesets.Sample</AssemblyTitle>
-    <OutputType>Library</OutputType>
-    <PlatformTarget>AnyCPU</PlatformTarget>
-    <RootNamespace>osu.Game.Rulesets.Tau</RootNamespace>
-    <LangVersion>8</LangVersion>
-  </PropertyGroup>
-  <ItemGroup>
-    <EmbeddedResource Include="Resources\**" />
-  </ItemGroup>
-  <ItemGroup>
-    <PackageReference Include="ppy.osu.Framework" Version="2020.403.0" />
-    <PackageReference Include="ppy.osu.Game" Version="2020.404.0" />
-  </ItemGroup>
-  <ItemGroup>
-    <Folder Include="Resources\Samples\Gameplay" />
-    <Folder Include="Resources\Textures" />
-  </ItemGroup>
+	<PropertyGroup Label="Project">
+		<TargetFramework>netcoreapp3.0</TargetFramework>
+		<AssemblyTitle>osu.Game.Rulesets.Sample</AssemblyTitle>
+		<OutputType>Library</OutputType>
+		<PlatformTarget>AnyCPU</PlatformTarget>
+		<RootNamespace>osu.Game.Rulesets.Tau</RootNamespace>
+		<LangVersion>8</LangVersion>
+	</PropertyGroup>
+	<ItemGroup>
+		<EmbeddedResource Include="Resources\**" />
+	</ItemGroup>
+	<ItemGroup>
+		<PackageReference Include="ppy.osu.Framework" Version="2020.403.0" />
+		<PackageReference Include="ppy.osu.Game" Version="2020.404.0" />
+	</ItemGroup>
+	<ItemGroup>
+		<Folder Include="Resources\Samples\Gameplay" />
+		<Folder Include="Resources\Textures" />
+	</ItemGroup>
 </Project>
